refactor: Split monolithic ui.rs into focused submodules
Break the 3619-line ui.rs into 7 focused submodules under ui/:
- mod.rs: Top-level layout, repo status, and module re-exports (~170 lines)
- board.rs: Kanban task swimlanes and coworker status table (~300 lines)
- chat.rs: Message display, input bar, and autocomplete dropdown (~250 lines)
- messages.rs: Message rendering (sender headers, timestamps, content) (~350 lines)
- messages_mermaid.rs: Mermaid diagram rendering within messages (~300 lines)
- styles.rs: Shared color and sender classification helpers (~70 lines)
- text.rs: Markdown parsing and line wrapping utilities (~180 lines)
- usage.rs: Usage progress bars with time estimation (~230 lines)

All 69 existing tests preserved and passing, co-located with their
respective modules. No functional changes.
